Fix CodeRabbit findings: extract sonar module, retire alias scripts, split test files #56
